requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.0.5
django-allauth==0.45.0django-cors-headers==3.7.0
django-filter==2.4.0
django-jsonfield==1.4.1django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.12.4djangorestframework-simplejwt==4.6.0
httplib2==0.19.1
idna==3.2oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.2.0pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
